requirements.txt
# LightRAG
lightrag-hku

# Microsoft GraphRAG  (also needed by src/compare.py)
graphrag

# Azure OpenAI SDK
openai

# HTTP requests (used directly for rate-limit control)
requests

# Numerical arrays (embeddings)
numpy

# Environment variable loading
python-dotenv

# Vector store (used by both LightRAG and GraphRAG)
lancedb
